third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-text/white-space/seg-break-transformation-004.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] linebreak only ₩２４
  assert_equals: expected 60 but got 67
[FAIL] spaces linebreak ₩２４
  assert_equals: expected 60 but got 67
[FAIL] linebreak spaces ₩２４
  assert_equals: expected 60 but got 67
[FAIL] spaces linebreak spaces ₩２４
  assert_equals: expected 60 but got 67
[FAIL] multiple linebreaks ₩２４
  assert_equals: expected 60 but got 67
[FAIL] multiple linebreaks + spaces ₩２４
  assert_equals: expected 60 but got 67
[FAIL] linebreak only ２４₩
  assert_equals: expected 60 but got 66
[FAIL] spaces linebreak ２４₩
  assert_equals: expected 60 but got 66
[FAIL] linebreak spaces ２４₩
  assert_equals: expected 60 but got 66
[FAIL] spaces linebreak spaces ２４₩
  assert_equals: expected 60 but got 66
[FAIL] multiple linebreaks ２４₩
  assert_equals: expected 60 but got 66
[FAIL] multiple linebreaks + spaces ２４₩
  assert_equals: expected 60 but got 66
Harness: the test ran to completion.
